+++ src/AST/Expresiones/Nativas/Hist.java
@@ -43,7 +43,7 @@ public class Hist extends Sentencia{
             LinkedList<Object> temp = (LinkedList)Odatos;
             for(Object element : temp){
                 Object tipo = Primitivo.getTipoDato(element);
-                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                     entorno.addError(new Token("Hist-TipoDeDato Parameter(x)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                     return null;
                 }
@@ -55,7 +55,7 @@ public class Hist extends Sentencia{
             }
         }else{
             Object tipo = Primitivo.getTipoDato(Odatos);
-            if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+            if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                 entorno.addError(new Token("Hist-TipoDeDato Parameter(x)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                 return null;
             }
