+++ src/AST/Expresiones/Nativas/Pie.java
@@ -44,7 +44,7 @@ public class Pie extends Sentencia{
             LinkedList<Object> temp = (LinkedList)Odatos;
             for(Object element : temp){
                 Object tipo = Primitivo.getTipoDato(element);
-                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                     entorno.addError(new Token("Pie-TipoDeDato Parameter(x)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                     return null;
                 }
@@ -56,7 +56,7 @@ public class Pie extends Sentencia{
             }
         }else{
             Object tipo = Primitivo.getTipoDato(Odatos);
-            if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+            if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                 entorno.addError(new Token("Pie-TipoDeDato Parameter(x)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                 return null;
             }
