+++ src/AST/Expresiones/Nativas/Plot.java
@@ -53,7 +53,7 @@ public class Plot extends Sentencia{
             
             for(Object element : temp){
                 Object tipo = Primitivo.getTipoDato(element);
-                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                     entorno.addError(new Token("Grafica-Plot-TipoDeDato Parameter(x)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                     return null;
                 }
@@ -79,7 +79,7 @@ public class Plot extends Sentencia{
             }
             for(Object element : temp){
                 Object tipo = Primitivo.getTipoDato(element);
-                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.DOUBLE))){
+                if(!(tipo.equals(Expresion.TIPO_PRIMITIVO.INTEGER) || tipo.equals(Expresion.TIPO_PRIMITIVO.NUMERIC))){
                     entorno.addError(new Token("Grafica-Plot-TipoDeDato Parameter(Ylim)", "Se esperaba un vector int o double, viene: "+tipo, fila, columna));
                     return null;
                 }
